encoder: E3, E4 and E5 on the GPU, byte-identical to the reference
E3 forward.comp is one workgroup of 64 lanes per tile: block means, the DC
plane quantised and then reconstructed the way the decoder will reconstruct
it, and the residual blocks eight at a time through an LDS transpose.  The
bilinear prediction is never materialised -- pred_at recomputes it from the 64
means, which is bit-identical and saves the 16 KB a 64x64 plane would cost.
Directional intra is behind a specialization constant that also sizes the
running reconstruction's shared array, so a build without it pays nothing;
the per-block modes are an input, because choosing them is a floating-point
search and belongs to the host or E1.

E4 rans_encode.comp is eight lanes per tile, eight tiles per workgroup,
persistent.  Two things there are not obvious and are written out in
rans_cpu.h and the shader head: a lane machine reaches kDone once, so the
global operation order is (round ascending, lane ascending) and the backward
sweep is eight lanes in lockstep over descending rounds with no list; and the
renormalisation words are anchored at the end of the tile's slot, so an
emission's position is known when it is produced and one sweep does the work
of two.

E5 packetize.comp lays the tiles out over E2's prefix sum, writing the frame
and tile-row headers itself, straight into the host-cached buffer.

The acid test: nxvc-vkenc and nxv-enc driven from one description of the same
picture must produce byte-identical streams, and nxv-dec must decode both to
identical pixels.  That holds over 32 configurations -- QP 0 to 63, all four
frame matrices, wm_id 1..3, nsub_log2 0..3, transform skip, both context
models, sign hiding on and off, chroma QP offsets, 4:2:0 and 4:4:4,
non-multiple-of-64 frames, side-by-side stereo, and 2 x 2048^2 -- on RADV and
on lavapipe, eight times apart in subgroup width.  Directional intra has no
reference stream to match, since nxv-enc searches its own modes; it is covered
against the CPU models, against pinned digests, and by the fact that mode 0
everywhere decodes to exactly what the v1 predictor produces.

Timing, 2 x 2048^2 (2048 tiles), RX 7900 XTX on RADV, median of 50: at QP 24
E3 0.92 ms, E4 2.22 ms, E2 0.01 ms, E5 0.10 ms, 3.25 ms total.  That is three
times the 7900 XTX budget of 3.6, and the reason is structural: E4 has exactly
tiles x 8 lanes of parallelism because eight rANS lanes per tile is the
bitstream, so it is latency-bound at under three waves per SIMD.  Three
changes took it from 29.5 ms to 2.2 ms without touching the inner loop --
every tile resident, each lane's operation list materialised once with an
exact fallback, and the single sweep.  The remaining lever, parallelising
generation over coding units rather than lanes, is named in the README rather
than claimed.

+++ vk/encoder/forward/rans_cpu.h
@@ -83,9 +83,11 @@ typedef struct nxe_tile_units {
  * exclusive cumulative frequencies cum[set][ctx][sym].  This is exactly the
  * buffer the shader binds; `cum` has NXE_NUM_SYM entries per context, not
  * NXE_NUM_SYM+1, because the encoder never needs the final total. */
+/* 32-bit entries, not 16: this struct is uploaded verbatim as the shader's
+ * table buffer, and a std430 uint array is what E4 indexes. */
 typedef struct nxe_tables {
-    uint16_t freq[8][NXE_MAX_CTX][NXE_NUM_SYM];
-    uint16_t cum[8][NXE_MAX_CTX][NXE_NUM_SYM];
+    uint32_t freq[8][NXE_MAX_CTX][NXE_NUM_SYM];
+    uint32_t cum[8][NXE_MAX_CTX][NXE_NUM_SYM];
 } nxe_tables;
 
 /* Build the tile's unit list, mirroring TileCoder::build_units. */
